requirements.txt
Django>=4.2.6,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2
drf-spectacular>=0.26.5,<0.27
